+++ lib/output/mqtt.go
@@ -33,7 +33,11 @@ func init() {
 	Constructors[TypeMQTT] = TypeSpec{
 		constructor: NewMQTT,
 		description: `
-Pushes messages to an MQTT broker.`,
+Pushes messages to an MQTT broker.
+
+The ` + "`topic`" + ` field can be dynamically set using function interpolations
+described [here](../config_interpolation.md#functions). When sending batched
+messages these interpolations are performed per message part.`,
 	}
 }