+++ CHANGELOG.md
@@ -13,6 +13,10 @@ All notable changes to this project will be documented in this file.
 - New `resource` processor.
 - Processors can now be registered within the `resources` section of a config.
 
+### Changed
+
+- The `mqtt` output field `topic` field now supports interpolation functions.
+
 ### Fixed
 
 - The `kafka` output no longer attempts to send headers on old versions of the